requirements.txt
sentence-transformers
faiss-cpu
numpy
streamlit
python-docx
openai
requests